agents/orchestrator_agent/pyproject.toml
[project]
name = "pet-orchestrator-agent"
version = "0.1.0"
description = "Pet Care Orchestrator Agent with A2A Protocol"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.5,<0.4.0",
    "httpx>=0.28.1",
    "langchain-openai",
    "langgraph",
    "langchain-core",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "starlette",
    "authlib>=1.3.0",  # OAuth 2.0 client for Asgardeo authentication
]

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
